# docs/openapi.yaml
openapi: 3.1.0
info:
  title: archimap API
  version: 1.0.0
  description: Canonical HTTP contract for ArchiMap. `docs/api.md` is the human-readable companion and should stay aligned with this file.
servers:
  - url: http://localhost:3252
paths:
...
  /api/building/{osmType}/{osmId}:
    get:
      summary: Get building GeoJSON feature by OSM id
      parameters:
        - in: path
          name: osmType
          required: true
          schema:
            type: string
            enum: [way, relation]
        - in: path
          name: osmId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Building feature
        '400':
          description: Invalid identifier
        '404':
          description: Building not found
  /api/building-info/{osmType}/{osmId}:
    get:
      summary: Get merged building information by OSM id, including design project fields and cached design-ref suggestions
      parameters:
        - in: path
          name: osmType
          required: true
          schema:
            type: string
            enum: [way, relation]
        - in: path
          name: osmId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Building info
        '400':
          description: Invalid identifier
        '404':
          description: Building info not found
  /api/building-info:
    post:
      summary: Submit or update a user building edit
      parameters:
        - in: header
          name: x-csrf-token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [osmType, osmId]
              properties:
                osmType:
                  type: string
                  enum: [way, relation]
                osmId:
                  type: integer
                name:
                  type: string
                style:
                  type: string
                design:
                  type: string
                designRef:
                  type: string
                designYear:
                  type: integer
                roofShape:
                  type: string
                levels:
                  type: integer
                yearBuilt:
                  type: integer
                architect:
                  type: string
                address:
                  type: string
                archimapDescription:
                  type: string
                editedFields:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Edit accepted
        '400':
          description: Invalid payload or building identifier
        '401':
          description: Unauthorized
        '403':
          description: Forbidden or CSRF failed
        '409':
          description: No effective changes to submit
...
  /api/buildings/filter-data-bbox:
    get:
      summary: Get filter payload for visible bbox
      parameters:
        - in: query
          name: minLon
          required: true
          schema:
            type: number
        - in: query
          name: minLat
          required: true
          schema:
            type: number
        - in: query
          name: maxLon
          required: true
          schema:
            type: number
        - in: query
          name: maxLat
          required: true
          schema:
            type: number
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Filter payload; items may include `centerLon` and `centerLat` for low-zoom marker fallback
        '400':
          description: Invalid bbox
  /api/buildings/filter-matches:
    post:
      summary: Match filter rules against bbox
      description: |
        The endpoint accepts a flat `rules[]` array per request.
        Client-side filter layers, priorities, colors, and AND/OR/layer composition are resolved before calling this endpoint.
        The response includes `matchedLocations[]` points for low-zoom marker fallback, in addition to `matchedFeatureIds[]` for contour highlight.
        When `renderMode=markers` and the zoom is below `5`, the backend can aggregate matches into viewport-relative cells and include `count` on each returned location; zooms `5-12` keep stable point-level locations.
        Use `POST /api/buildings/filter-matches-batch` when you need multiple rule sets against the same bbox.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bbox, rules]
              properties:
                bbox:
                  type: object
                  properties:
                    west:
                      type: number
                    south:
                      type: number
                    east:
                      type: number
                    north:
                      type: number
                zoom:
                  type: number
                zoomBucket:
                  type: number
                renderMode:
                  type: string
                  enum: [contours, markers]
                maxResults:
                  type: integer
                rules:
                  type: array
                  description: Flat rule list for one request. Supported operators are `contains`, `equals`, `not_equals`, `starts_with`, `exists`, `not_exists`, `greater_than`, `greater_or_equals`, `less_than`, and `less_or_equals`.
                  items:
                    type: object
                    properties:
                      key:
                        type: string
                      op:
                        type: string
                        enum:
                          [
                            contains,
                            equals,
                            not_equals,
                            starts_with,
                            exists,
                            not_exists,
                            greater_than,
                            greater_or_equals,
                            less_than,
                            less_or_equals
                          ]
                        description: '`exists` and `not_exists` do not require `value`. Numeric operators expect a numeric string in `value`.'
                      value:
                        type: string
                        description: Optional for `exists` and `not_exists`.
      responses:
        '200':
          description: Match result
        '400':
          description: Invalid request
        '429':
          description: Rate limited
...
  /api/buildings/filter-data:
    post:
      summary: Get merged building filter data for explicit keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keys]
              properties:
                keys:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Filter data payload; items may include `centerLon` and `centerLat` for low-zoom marker fallback
        '400':
          description: Invalid keys payload
...
  /api/buildings/filter-matches-batch:
    post:
      summary: Match multiple filter rule sets against one bbox
      description: |
        Batched variant of `POST /api/buildings/filter-matches`.
        Each item includes `matchedLocations[]` for low-zoom marker fallback, alongside `matchedFeatureIds[]` for contour highlight.
        When `renderMode=markers` and the zoom is below `5`, the backend can return aggregated marker cells with `count` instead of a full building list; zooms `5-12` keep stable point-level locations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bbox, requests]
              properties:
                bbox:
                  type: object
                  required: [west, south, east, north]
                  properties:
                    west:
                      type: number
                    south:
                      type: number
                    east:
                      type: number
                    north:
                      type: number
                zoom:
                  type: number
                zoomBucket:
                  type: number
                renderMode:
                  type: string
                  enum: [contours, markers]
                requests:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      rulesHash:
                        type: string
                      maxResults:
                        type: integer
                      rules:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            op:
                              type: string
                              enum:
                                [
                                  contains,
                                  equals,
                                  not_equals,
                                  starts_with,
                                  exists,
                                  not_exists,
                                  greater_than,
                                  greater_or_equals,
                                  less_than,
                                  less_or_equals
                                ]
                            value:
                              type: string
      responses:
        '200':
          description: Batch match result payload
        '400':
          description: Invalid batch request
        '429':
          description: Rate limited
...